fixed imports and packaging temporarily

+++ backend/src/main/java/com/crashcourse/kickoff/tms/club/ClubService.java
@@ -1,9 +1,13 @@
-package com.example.tournamentmanagement.service;
+package com.crashcourse.kickoff.tms.club;
 
-import com.example.tournamentmanagement.entity.Club;
-import com.example.tournamentmanagement.repository.ClubRepository;
+import com.crashcourse.kickoff.tms.club.exception.*;
+import com.crashcourse.kickoff.tms.user.User;
+import com.crashcourse.kickoff.tms.tournament.*;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,7 +17,21 @@ public class ClubService {
     @Autowired
     private ClubRepository clubRepository;
 
-    public Club createClub(Club club) {
+    // jparepository has automatically implemented crud methods
+    public Club createClub(@Valid Club club, User creator) {
+        // club name not unique
+        if (clubRepository.findByName(club.getName()).isPresent()) {
+            throw new ClubAlreadyExistsException("Club name must be unique");
+        }
+
+        // set the player who created the club as the captain
+        club.setCaptain(creator);
+
+        // player count exceeds the limit
+        if (club.getPlayers().size() > Club.MAX_PLAYERS_IN_CLUB) {
+            throw new PlayerLimitExceededException(String.format("A club cannot have more than %d players", Club.MAX_PLAYERS_IN_CLUB));
+        }
+
         return clubRepository.save(club);
     }
 
@@ -22,27 +40,41 @@ public class ClubService {
     }
 
     public Optional<Club> getClubById(Long id) {
-        return clubRepository.findById(id);
+        Optional<Club> club = clubRepository.findById(id);
+        if (!club.isPresent()) {
+            throw new ClubNotFoundException("Club with ID " + id + " not found");
+        }
+        return club;
     }
-    
+
     public void deleteClub(Long id) {
+        if (!clubRepository.existsById(id)) {
+            throw new ClubNotFoundException("Club with ID " + id + " not found");
+        }
         clubRepository.deleteById(id);
     }
 
-    // Update an existing club (can use this when you need to add or remove players)
+    // general method to update club, if there's common use case for a specific method (eg. updateElo), we can make that too
     public Club updateClub(Long id, Club clubDetails) {
         Optional<Club> clubOptional = clubRepository.findById(id);
         if (clubOptional.isPresent()) {
-            // if you have more club instance variables, call setter for all of them
             Club club = clubOptional.get();
+            
+            // new name is not unique
+            if (!club.getName().equals(clubDetails.getName()) &&
+                clubRepository.findByName(clubDetails.getName()).isPresent()) {
+                throw new ClubAlreadyExistsException("Club name must be unique");
+            }
+            
+            // call setters to change deets
             club.setName(clubDetails.getName());
             club.setElo(clubDetails.getElo());
             club.setRatingDeviation(clubDetails.getRatingDeviation());
-
+            
             return clubRepository.save(club);
         }
 
-        // if club doesn't exist, return null
-        return null;
+        // no such club to update
+        throw new ClubNotFoundException("Club with ID " + id + " not found");
     }
 }
